feat: add --static which makes all files of same size identical
By default, writedisk will write a byte buffer of sequential byte
values, starting with a _random_ value, and rolling to 0x00 when
the sequence passes 0xff.

When --static is used, the starting value is always 0x00. This means
that any file created with --static and of the same size will always
contain the exist same values. This is useful when testing APFS
cloning support, for example.